fixed some more references
