+++ texfiles/0.intro.tex
@@ -5,7 +5,6 @@
 
 \chapter{Computers, \mbox{Objects}, and Java}
 \label{chapter-intro}
-\label{pg-chapter-intro}
 
 \CObegin
 \secCOBH{Objectives}
@@ -79,8 +78,7 @@ components (Fig.~\ref{fig-blockdiagram}).
 {A diagram of the main functional
 components in a computer system. The arrows indicate the flow of information
 between various components.
-\label{fig-blockdiagram}
-\label{pg-fig-blockdiagram}}
+\label{fig-blockdiagram}}
 %\end{graphic}
 \end{figure}
 
@@ -301,8 +299,7 @@ through this (relatively) simple mechanism.
 {WWW: The client's browser requests a page from
 a Web server.  When the HTML document is returned, it is
 interpreted and displayed by the browser.
-\label{fig-http}
-\label{pg-fig-http}}
+\label{fig-http}}
 %\end{graphic}
 \end{figure}
 
@@ -641,11 +638,10 @@ interact with and communicate with each other.
 
 A Java program, and any object-oriented program, is a collection of
 interacting objects that models a collection of real-world
-\marginfig{chptr00/kitchen.eps} {A model of a kitchen.
-\label{fig-kitchen}
-\label{pg-fig-kitchen}}
+\marginfig{chptr00/kitchen.eps} {A model of a kitchen.}{fig-kitchen}
+
 objects. Think of the model that a kitchen designer might use to
-layout your new kitchen (Fig.~4).  It will contain
+layout your new kitchen (Fig.~\ref{fig-kitchen}).  It will contain
 objects that represent the various kitchen appliances and
 cabinets. Each object in the model is a simplified version of the
 corresponding real object.  For example, a rectangle might be used to
@@ -774,12 +770,11 @@ message does not involve an argument. But it does involve a result. A
 that sent the message.
 
 \begin{figure}[h]
-\figa{chptr00/atmreport.eps} {This UML diagram illustrates an ATM
+\figascaled{chptr00/atmreport.png}{0.5} {This UML diagram illustrates an ATM
 transaction in which a customer asks the ATM machine for his current
 balance. The ATM gets this information from an object representing the
-customer's bank account and passes it to the customer.
-\label{fig-umlmsg2}
-\label{pg-fig-umlmsg2}}
+customer's bank account and passes it to the customer.}
+{fig-umlmsg2}
 \end{figure}
 
 Obviously, in order to respond to a message, an object has to know how
@@ -966,9 +961,9 @@ its width.
 
 \begin{figure}[tb]
 \figa{chptr00/classhier.eps} 
-{A hierarchy of Java classes. 
-\label{fig-classhier}
-\label{pg-fig-classhier}}
+{A hierarchy of Java classes.}
+{fig-classhier}
+
 \end{figure}
 
 To introduce some important terminology associated with this kind of
@@ -1008,7 +1003,7 @@ column} attributes are given different values in each {\tt ChessPiece}
 object.
 
 \begin{figure}[tb]
-\figaleft{chptr00/chesshier.eps} 
+\figaleft{chptr00/chesshier-small.png} 
 {The {\tt ChessPiece} hierarchy.
 \label{fig-chesshier}
 \label{pg-fig-chesshier}}
