+++ texfiles/3.methods.tex
@@ -326,8 +326,8 @@ public class OneRowNim
 } //OneRowNim1 class
 \end{lstlisting}
 \end{jjjlisting}
-\jjjprogstop{The {\tt OneRowNim} class definition with {\tt takeSticks()} method.
-\label{fig-ornclass2}}
+\jjjprogstop{The {\tt OneRowNim} class definition with {\tt takeSticks()} method.}
+{fig-ornclass2}
 \end{figure}
 single method, {\tt takeSticks()}, that can be used to take away a
 variable number of sticks, we have removed the three methods we wrote
@@ -452,9 +452,8 @@ public static void main (String argv[])
 } // main()
 \end{lstlisting}
 \end{jjjlisting}
-\jjjprogstop{A {\tt main()} method to test the {\tt takeSticks()} method.
-\label{fig-testorn1}
-\label{pg-fig-testorn1}}
+\jjjprogstop{A {\tt main()} method to test the {\tt takeSticks()} method.}
+{fig-testorn1}
 \end{figure}
 of {\tt takeSticks()} interacts with the instance variables {\tt
 nSticks} and {\tt player}.  The statements in the {\tt main()} program
@@ -516,8 +515,8 @@ public static void main (String argv[])
 } // main()
 \end{lstlisting}
 \end{jjjlisting}
-\jjjprogstop{A {\tt main()} method with keyboard input for {\tt One\-Row\-Nim}.
-\label{fig-testorn2}}
+\jjjprogstop{A {\tt main()} method with keyboard input for {\tt One\-Row\-Nim}.}
+{fig-testorn2}
 \end{figure}
 this program the following output will be generated in the console
 window before waiting for keyboard input:
@@ -1077,8 +1076,8 @@ public static void main(String[] args)
 } //main()
 \end{lstlisting}
 \end{jjjlisting}
-\jjjprogstop{A {\tt main()} method that tests the new methods for {\tt OneRowNim}
-\label{fig-testorn3}}
+\jjjprogstop{A {\tt main()} method that tests the new methods for {\tt OneRowNim}}
+{fig-testorn3}
 \end{figure}
 
 \noindent The output to the console  when this program is run will
@@ -1156,9 +1155,8 @@ public class PrimitiveCall
 } // PrimitiveCall
 \end{lstlisting}
 \end{jjjlisting}
-\jjjprogstop{Passing a primitive value to a method.
-\label{fig-primcall}
-\label{pg-fig-primcall}}
+\jjjprogstop{Passing a primitive value to a method.}
+{fig-primcall}
 \end{figure}
 parameter.  For example, consider the \mbox{\tt PrimitiveCall} class
 shown in Figure~\ref{fig-primcall}. Note that we have an {\tt int}
@@ -1243,8 +1241,8 @@ public class ReferenceCall
 } // ReferenceCall
 \end{lstlisting}
 \end{jjjlisting}
-\jjjprogstop{Passing a reference value to a method.
-\label{fig-refcall}}
+\jjjprogstop{Passing a reference value to a method.}
+{fig-refcall}
 \end{figure}
 parameter {\tt g} of type {\tt OneRowNim}. Because a \mbox{\tt
 OneRowNim} instance is an object, {\tt g} is a reference variable.  So
@@ -1282,7 +1280,7 @@ in {\tt ReferenceCall} (a) Just
 before calling {\tt myMethod(game)}. (b) Just before executing the body
 of {\tt myMethod()}. (c) Just after executing the body of {\tt
 myMethod()}. (d) After flow of control returns to {\tt main()}.}
-\label{fig-rparameps}
+{fig-rparameps}
 \end{figure}
 
 \noindent This illustrates that when passing a reference variable
@@ -1405,8 +1403,8 @@ Figure~3.11. The flowchart clearly shows that program flow will take
 Diamond-shaped symbols at  the branch points contain {\tt boolean}
 expressions.  Rectangular  symbols can only contain executable
 statements.  Circles act  simply as connectors, to connect two
-or more paths.
-\label{fig-ifflowchart}}
+or more paths.}
+{fig-ifflowchart}
 one or the other of the alternative paths coming out of the
 diamond-shaped boolean condition box.  The branch through the
 rectangular statement box will be taken when the boolean condition is
@@ -1431,9 +1429,9 @@ public String getPlayerString()
 \noindent The flowchart in Figure~\ref{fig-getplayerflow} shows the program flow
 of the entire {\tt getPlayerString()} method.  It is important to note
 \begin{figure}[htb]
-\figascaled{chptr03/player.eps}{0.9}{Flowchart of the {\tt getPlayerString()} method.
+\figascaled{chptr03/player.eps}{0.9}{Flowchart of the {\tt getPlayerString()} method.}
 %%\figa{chptr03/player.eps}{Flowchart of the {\tt getPlayerString()} method.
-\label{fig-getplayerflow}}
+{fig-getplayerflow}
 \end{figure}
 that when a {\tt return}\index{return statement} statement is executed
 in a method, control is returned immediately to the calling method.
@@ -1551,9 +1549,9 @@ Semantically, the if-else statement\index{if-else statement} has the
 %\marginnote{If-else semantics}
 following interpretation: If the {\it boolean expression} is true,
 execute {\it statement1}; otherwise execute {\it statement2}.  This
-\marginfig{chptr03/3f14.eps}{Flowchart of the {\tt if-else} statement.
-\label{fig-ifelsechart}}
-interpretation is shown in Figure~3.13.
+\marginfig{chptr03/3f14.eps}{Flowchart of the {\tt if-else} statement.}
+{fig-ifelsechart}
+interpretation is shown in Figure~\ref{fig-ifelsechart}.
 
 \subsection{The Nested if/else Multiway Selection Structure}
 \label{sec-multi}
@@ -1987,8 +1985,8 @@ public class OneRowNim
 } // OneRowNim class
 \end{lstlisting}
 \end{jjjlisting}
-\jjjprogstop{The {\tt OneRowNim} class with improved methods.
-\label{fig-ornclass}}
+\jjjprogstop{The {\tt OneRowNim} class with improved methods.}
+{fig-ornclass}
 \end{figure}
 
 Let's use a while statement to test the new methods of the class.  A
@@ -2034,8 +2032,8 @@ public class TestOneRowNim
 } // TestOneRowNim
 \end{lstlisting}
 \end{jjjlisting}
-\jjjprogstop{The {\tt TestOneRowNim} class with a while loop.
-\label{fig-tornclass}}
+\jjjprogstop{The {\tt TestOneRowNim} class with a while loop.}
+{fig-tornclass}
 \end{figure}
 introduced in the previous chapter to get moves from the keyboard
 for both players.  Before each move {\tt game.report()} describes 
@@ -2262,8 +2260,8 @@ public class TestToString
 } //TestToString
 \end{lstlisting}
 \end{jjjlisting}
-\jjjprogstop{An application to test the overridden {\tt toString()} method.
-\label{fig-orntostring}}
+\jjjprogstop{An application to test the overridden {\tt toString()} method.}
+{fig-orntostring}
 \end{figure}
 
 \noindent While this new method may not play an important role in the
@@ -2468,8 +2466,8 @@ public class DrawSticksApplet extends Applet
 } // DrawSticksApplet
 \end{lstlisting}
 \end{jjjlisting}
-\jjjprogstop{An applet program with a method for drawing a set of sticks.
-\label{fig-drawsticksprog}}
+\jjjprogstop{An applet program with a method for drawing a set of sticks.}
+{fig-drawsticksprog}
 \end{figure}
 
 Note that the body of {\tt drawSticks()} uses a while-loop to draw the lines
@@ -2709,7 +2707,7 @@ public boolean playerOneIsNext()
 
 %\begin{figure}
 \marginfiglarge{chptr03/isdone.eps}{Flowchart of the if-else version of the
-{\tt getStatus()} method. \label{fig-getstateifelse}} 
+{\tt getStatus()} method.} {fig-getstateifelse}
 %\end{figure}
