+++ texfiles/4.uis.tex
@@ -5,7 +5,7 @@
 \setcounter{SSTUDYcount}{1}
 \chapter{Input/Output: Designing the User Interface}
 \label{chapter-io}
-\label{pg-chapter-io}
+
 
 \CObegin
 \secCOBH{Objectives}
@@ -117,9 +117,8 @@ console (Fig.~\ref{fig-command-line}).
 
 \begin{figure}[h]
 \figa{chptr04/cmdline.eps}
-{A command-line user interface.
-\label{fig-command-line}
-\label{pg-fig-command-line}}
+{A command-line user interface.}
+{fig-command-line}
 \end{figure}
 
 The command-line approach might also be called {\em console
@@ -313,9 +312,9 @@ involved in performing keyboard input.
 
 \begin{figure}[h]
 \figa{chptr04/keybrdrdr.eps}
-{Design of the {\tt KeyboardReader} class.
-\label{fig-keybrdrdr}
-\label{pg-fig-keybrdrdr}}
+{Design of the {\tt KeyboardReader} class.}
+{fig-keybrdrdr}
+
 \end{figure}
 
 Figure~\ref{fig-keybrdrdr} presents the design of {\tt KeyboardReader}
@@ -420,9 +419,8 @@ public class KeyboardReader
 }
 \end{lstlisting}
 \end{jjjlisting}
-\jjjprogstop{Definition of the {\tt Keyboard\-Reader} class.
-\label{fig-keybrdrdrclass}
-\label{pg-fig-keybrdrdrclass}}
+\jjjprogstop{Definition of the {\tt Keyboard\-Reader} class.}
+{fig-keybrdrdrclass}
 \end{figure}
 line. The first thing to notice is the use of the {\tt import}
 statement. Recall that importing a Java package enables us to refer to
@@ -531,9 +529,8 @@ will handle the task of greeting the user.
 
 \begin{figure}[h]
 \figascaled{chptr04/greeter.eps}{0.9}
-{Using {\tt Key\-board\-Read\-er} as the user interface.
-\label{fig-greeter}
-\label{pg-fig-greeter}}
+{Using {\tt Key\-board\-Read\-er} as the user interface.}
+{fig-greeter}
 \end{figure}
 
 The full implementation of the {\tt GreeterApp} class is shown in
@@ -564,9 +561,8 @@ public class GreeterApp
 } // GreaterApp
 \end{lstlisting}
 \end{jjjlisting}
-\jjjprogstop{Definition of the {\tt GreeterApp} class.
-\label{fig-greeterapp}
-\label{pg-fig-greeterapp}}
+\jjjprogstop{Definition of the {\tt GreeterApp} class.}
+{fig-greeterapp}
 \end{figure}
 variable for the {\tt KeyboardReader}, which is instantiated in the
 constructor method. This gives {\tt GreeterApp} a way to refer
@@ -688,9 +684,9 @@ components.
 \figa{chptr04/guicapture.eps}
 %%%\figaleft{chptr04/guicapture.eps}
 {Various GUI components from the {\tt javax.swing} package. {\bf [Artwork: We
-need to label the components.]}
-\label{fig-guiscreen}
-\label{pg-fig-guiscreen}}
+need to label the components.]}}
+{fig-guiscreen}
+
 \end{figure}
 
 The Swing classes are generally considered to be superior to their AWT
@@ -868,8 +864,8 @@ public class SimpleGUI extends JFrame
 } // SimpleGUI class
 \end{lstlisting}
 \end{jjjlisting}
-\jjjprogstop{A top-level window with a title.
-\label{fig-simpleframe}}
+\jjjprogstop{A top-level window with a title.}
+{fig-simpleframe}
 \end{figure}
 keyword to define {\tt SimpleGUI} as a subclass of {\tt JFrame}.  As a
 subclass, {\tt SimpleGUI} inherits all of the functionality of a {\tt
@@ -897,9 +893,9 @@ done in the constructor method.
 \begin{figure}[h!]
 \figa{chptr04/framesubclass.eps}
 %%%\figaleft{chptr04/framesubclass.eps}
-{{\tt SimpleGUI} is a subclass of {\tt JFrame}.
-\label{fig-framesubclass}
-\label{pg-fig-framesubclass}}
+{{\tt SimpleGUI} is a subclass of {\tt JFrame}.}
+{fig-framesubclass}
+
 \end{figure}
 
 The constructor method illustrates how to use some of the methods
@@ -1099,9 +1095,9 @@ software, such as a browser or a word processor.
 \begin{figure}[tb]
 \figa{chptr04/eventmodel.eps}   %%%RAM See figure 4-15, p 185 in 2E
 %%%\figaleft{chptr04/eventmodel.eps}
-{Java's event model.
-\label{fig-eventmodel}
-\label{pg-fig-eventmodel}}
+{Java's event model.}
+{fig-eventmodel}
+
 \end{figure}
 
 In an event-driven programming model, the program is controlled by
@@ -1363,9 +1359,8 @@ public class GreeterGUI extends JFrame
 }
 \end{lstlisting}
 \end{jjjlisting}
-\jjjprogstop{Definition of the {\tt GreeterGUI} class.
-\label{fig-simplegui}
-\label{pg-simplegui}}
+\jjjprogstop{Definition of the {\tt GreeterGUI} class.}
+{fig-simplegui}
 \end{figure}
 even though we have introduced most of its elements already. That
 will help us put  together all of the various concepts that
@@ -1553,9 +1548,8 @@ public class GreeterGUIPanel extends JPanel
 } // GreeterGUIPanel
 \end{lstlisting}
 \end{jjjlisting}
-\jjjprogstop{Definition of the {\tt GreeterGUIPanel} class.
-\label{fig-guipanel}
-\label{pg-guipanel}}
+\jjjprogstop{Definition of the {\tt GreeterGUIPanel} class.}
+{fig-guipanel}
 \end{figure}
 
 The revised {\tt GreeterGUIPanel} class requires the following two
@@ -1617,8 +1611,8 @@ problem into two primary objects: a computational object, in this case
 use either a {\tt KeyboardReader} or a {\tt OneRowNimGUI}. One goal of
 our design was to develop the {\tt OneRowNim} class so that it can be
 used, without changes, with either a command-line interface or a GUI.
-\marginfig{chptr04/ornuml.eps}{A UML diagram of the {\tt OneRowNim} class.
-\label{fig-ch4ornuml}}
+\marginfig{chptr04/ornuml.eps}{A UML diagram of the {\tt OneRowNim} class.}
+{fig-ch4ornuml}
 
 Recall that we designed the {\tt OneRowNim} class to maintain the state of
 the game and to provide methods that enforce the rules of the
@@ -1789,8 +1783,8 @@ public class OneRowNimApp
 \end{lstlisting}
 \end{jjjlisting}
 \jjjprogstop{Definition of {\tt OneRowNimApp}, a command-line interface
-to the {\tt OneRowNim}.
-\label{fig-ornapp}}
+to the {\tt OneRowNim}.}
+{fig-ornapp}
 \end{figure}
 
 
@@ -1969,8 +1963,8 @@ public class OneRowNimGUI extends JFrame implements ActionListener
 } // OneRowNimGUI
 \end{lstlisting}
 \end{jjjlisting}
-\jjjprogstop{The {\tt OneRowNimGUI} class.
-\label{fig-orngui}}
+\jjjprogstop{The {\tt OneRowNimGUI} class.}
+{fig-orngui}
 \end{figure}
 
 Figure~\ref{fig-orngui} shows the complete implementation of the
@@ -2088,8 +2082,8 @@ file created by standard text editors like {\it NotePad} and {\it WordPad}
 on a Windows computer or {\it SimpleText} on a Macintosh. A {\bf binary file}
 has a more general format that can store numbers and other data the way they are
 \marginfig{chptr04/fileuml.eps} {A UML class diagram of the {\tt File}
- class with a partial list of public methods 
- \label{fig-fileuml}}
+ class with a partial list of public methods }
+{fig-fileuml}
 stored in the computer.  In this section we will consider only text files. Binary
 files are considered in Chapter~11.
 
@@ -2112,8 +2106,8 @@ statement
 
 \marginfig{chptr04/scanner2uml.eps}
 {A UML class diagram of the {\tt Scanner} class with an
-expanded list of public methods
-\label{fig-scanner2uml}}
+expanded list of public methods}
+{fig-scanner2uml}
 
 \noindent will obtain and store information about the "riddles.txt" file in
 the same directory as the java code being executed, if such a file
@@ -2223,8 +2217,8 @@ public class RiddleFileReader
 }  //RiddleFileReader class
 \end{lstlisting}
 \end{jjjlisting}
-\jjjprogstop{A program which reads riddles from a file and displays them.
-\label{fig-fileprog}}
+\jjjprogstop{A program which reads riddles from a file and displays them.}
+{fig-fileprog}
 \end{figure}
 and displays them.  Note that, in the body of the method {\tt readRiddles()},
 the statements:
