+++ texfiles/5.data.tex
@@ -602,9 +602,8 @@ public class OneRowNim
 \end{lstlisting}
 \end{jjjlisting}
 \jjjprogstop{The revised {\tt One\-Row\-Nim} uses a {\tt boolean}
-variable to keep track of who's turn it is.
-\label{fig-newonerow}
-\label{pg-fig-newonerow}}
+variable to keep track of who's turn it is.}
+{fig-newonerow}
 \end{figure}
 the use of {\tt boolean} variables. The {\tt takeSticks()} method
 contains two revisions. The first uses the boolean OR operator to test
@@ -1781,10 +1780,9 @@ floating-point literals (9.0, 5.0, and 32.0) in its conversion
 expression. This helps to reduce the reliance on Java's built-in
 promotion rules, which can lead to subtle errors.  For example,
 \marginfig{chptr05/p242f2.eps}{The {\tt Temperature} class. Note that static 
-elements are underlined in UML.  (NEEDS REVISION)
-\label{fig-p242f2}
-\label{pg-fig-p242f2}}
-to the design shown in Figure~5.4.
+elements are underlined in UML.  (NEEDS REVISION)}
+{fig-p242f2}
+to the design shown in Figure~\ref{fig-p242f2},
 suppose we had written what looks like an equivalent expression using
 integer literals:
 
@@ -1814,9 +1812,8 @@ public class Temperature
 } // Temperature
 \end{lstlisting}
 \end{jjjlisting}
-\jjjprogstop{The {\tt Temperature} class.
-\label{fig-temperature}
-\label{pg-fig-temperature}}
+\jjjprogstop{The {\tt Temperature} class.}
+{fig-temperature}
 \end{figure}
 
 \noindent Because 9 divided by 5 gives the integer result 1, this
@@ -1890,9 +1887,8 @@ handle interaction with the user and uses {\tt static} methods in the
 {\tt Temperature} class to perform the temperature conversions.
 
 \begin{figure}[h!]
-\figa{chptr05/tempui.eps}{A command-line user interface.
-\label{fig-tempcmdline}
-\label{pg-fig-tempcmdline}}
+\figa{chptr05/tempui.eps}{A command-line user interface.}
+{fig-tempcmdline}
 \end{figure}
 
 
@@ -1916,16 +1912,15 @@ double fahr = Temperature.celsToFahr(98.6);
 
 \item  Following the design for the GUI developed in Chapter~4,
 \marginfig{chptr05/5f5.eps}{Layout design of a GUI that
-performs temperature conversions.
-\label{fig-convapplgui}
-\label{pg-fig-convapplgui}}
+performs temperature conversions.}
+{fig-convapplgui}
 implement a GUI to use for testing the {\tt Temperature} class.
-The GUI should have the layout shown in Figure~5.7.
+The GUI should have the layout shown in Figure~\ref{fig-convapplgui}.
 \end{SSTUDY}
 
 \subsection{Example: Using Class Constants}
 
-\noindent As we noted in Chapter~0, in
+\noindent As we noted in Chapter~\ref{chapter-intro}, in
 addition to instance variables, which are associated with instances
 (objects) of a class, Java also allows class variables, which are
 associated with the class itself.  One of the most common uses of such
@@ -2129,9 +2124,8 @@ public class OneRowNim
 } // OneRowNim class
 \end{lstlisting}
 \end{jjjlisting}
-\jjjprogstop{This version of {\tt One\-Row\-Nim} uses named constants.
-\label{fig-constnim}
-\label{pg-fig-constnim}}
+\jjjprogstop{This version of {\tt One\-Row\-Nim} uses named constants.}
+{fig-constnim}
 \end{figure}
 
 \secEXRHone{Self-Study Exercise}
@@ -2404,9 +2398,8 @@ expression
 the game at that point.
 
 \secEXRHone{Self-Study Exercise}
-\marginfig{chptr05/nimplayer.eps}{The {\tt NimPlayer} class.
-\label{fig-nimplayer}
-\label{pg-fig-nimplayer}}
+\marginfig{chptr05/nimplayer.eps}{The {\tt NimPlayer} class.}
+{fig-nimplayer}
 \begin{SSTUDY}
 
 \item  Implement a class named {\tt NimPlayer} that
@@ -2432,9 +2425,8 @@ computer.
 \tBOXseven{Although} the {\tt Math.round()} method is useful for rounding
 %\begin{figure}
 %\begin{graphic}
-\marginfig{chptr05/p267f1.eps}{The {\tt java.text.NumberFormat} class.
-\label{fig-p267f1}
-\label{pg-fig-p267f1}}
+\marginfig{chptr05/p267f1.eps}{The {\tt java.text.NumberFormat} class.}
+{fig-p267f1}
 numbers, it is not suitable for business applications. Even
 for rounded values, Java will drop trailing zeroes. So a value
 such as \$10,000.00 would be output as \$10000.0. This wouldn't
@@ -2442,7 +2434,7 @@ be acceptable for a business report.
 
 Fortunately, Java supplies the {\tt java.text.NumberFormat}
 class precisely for the task of representing numbers as dollar
-amounts, percentages, and other formats (Fig.~5.10).
+amounts, percentages, and other formats (Fig.~\ref{fig-pg267f1}).
 
 The {\tt NumberFormat} class is an {\tt abstract} class, which means
 that it cannot be directly instantiated. Instead, you would use its
@@ -2493,11 +2485,10 @@ Java class library.}
 
 \item A Certificate of Deposit (CD) is an investment instrument
 that accumulates interest at a given rate for an initial principal
-\marginfig{chptr05/bankcd.eps}{The {\tt BankCD} class.
-\label{fig-bankcd}
-\label{pg-fig-bankcd}}
+\marginfig{chptr05/bankcd.eps}{The {\tt BankCD} class.}
+{fig-bankcd}
 over a fixed number of years.  The formula for compounding interest is
-shown in Table~5.12. It assumes that interest is compounded
+shown in Table~\ref{fig-bankcd}. It assumes that interest is compounded
 annually. For daily compounding, the annual rate must be divided by
 365, and the compounding period must be multiplied by 365, giving: $a
 = p(1 + r/365)^{365n}$.  Implement a {\tt BankCD} class that
@@ -3041,8 +3032,8 @@ public class Test {
 \end{jjjlisting}
 \jjjprogstop{A Java program illustrating character\index{character
 conversion example} conversions. When run, the program will generate
-the following outputs, one per line: a, 98, b, A, B, 7.
-\label{fig-testapplet}\label{pg-fig-testapplet}}
+the following outputs, one per line: a, 98, b, A, B, 7.}
+{fig-testapplet}
 \end{figure}
 
 \section{Problem Solving = Representation + Action}
