+++ texfiles/6.loops.tex
@@ -262,8 +262,8 @@ process is repeated until the loop entry condition becomes false.
 
 \begin{figure}[h!]
 \figa{chptr03/whileflow.eps}{Flowchart of the {\tt while} statement and {\tt while}
-structure.
-\label{fig-whileflow6}}
+structure.}
+{fig-whileflow6}
 \end{figure}
 
 Tracing the order in which the {\tt for} loop components are
@@ -875,9 +875,8 @@ public class CarLoan {
 } // CarLoan
 \end{lstlisting}
 \end{jjjlistingleft}
-\jjjprogstop{The {\tt CarLoan} application.
-\label{fig-carloan}
-\label{pg-fig-carloan}}
+\jjjprogstop{The {\tt CarLoan} application.}
+{fig-carloan}
 \end{figure}
 
 \newpage
@@ -890,14 +889,13 @@ concentrate on drawing the checkerboard.  We will not worry about
 providing a full checkerboard representation, of the sort we would
 need if we were writing a program to play the game of checkers. So,
 our design will not involve instance
-\marginfig{chptr06/checkerboard.eps}{A checkerboard with checkers.
-\label{fig-checkerboard}
-\label{pg-fig-checkerboard}}
+\marginfig{chptr06/checkerboard.eps}{A checkerboard with checkers.}
+{fig-checkerboard}
 variables for whose turn it is, how many checkers each player has,
 where the checkers are located on the board, and other elements of the
 game's state. Still, our visible representation of the board should
 be designed so that it will eventually be useful when we do develop a
-checkers game in Chapter~8.
+checkers game in Chapter~\ref{chapter-inheritance}.
 
 \subsection*{Problem Description and Specification}
 
@@ -933,9 +931,8 @@ public class CheckerBoardApplet extends JApplet {
 } // CheckerBoardApplet
 \end{lstlisting}
 \end{jjjlisting}
-\jjjprogstop{The {\tt Chek\-er\-Board\-Applet} class.
-\label{fig-checkerapplet}
-\label{pg-fig-checkerapplet}}
+\jjjprogstop{The {\tt Chek\-er\-Board\-Applet} class.}
+{fig-checkerapplet}
 \end{figure}
 shown there, the applet simply creates a {\tt CheckerBoard} instance
 in its {\tt init()} method, and then invokes the {\tt CheckerBoard}'s
@@ -951,9 +948,8 @@ checkerboard.
 
 \subsection*{Class Design: {\tt CheckerBoard}}
 \marginfig{chptr06/checkeruml.eps}{Design of the {\tt CheckerBoard}
-class.
-\label{fig-checkeruml}
-\label{pg-fig-checkeruml}}
+class.}
+{fig-checkeruml}
 \noindent Because the applet will invoke its {\tt draw()} method, this
 method must be part of the {\tt CheckerBoard}'s interface.  Hence, it
 must be declared {\tt public}.  The task of drawing a checkerboard
@@ -965,7 +961,7 @@ methods that encapsulate these two subtasks. This is good method
 design because it results in relatively small methods, each of which
 performs a very well-defined task.  Let's call these methods {\tt
 drawBoard()} and {\tt drawCheckers()}, respectively.  Their signatures
-are shown in Figure~6.6, which summarizes the design of the {\tt
+are shown in Figure~\ref{fig-checkeruml}, which summarizes the design of the {\tt
 CheckerBoard} class.
 
 Before gettinginto the details of the {\tt drawBoard} and {\tt
@@ -1055,9 +1051,8 @@ public class CheckerBoard {
 } //CheckerBoard
 \end{lstlisting}
 \end{jjjlisting}
-\jjjprogstop{The {\tt CheckerBoard} class.
-\label{fig-checkerclass}
-\label{pg-fig-checkerclass}}
+\jjjprogstop{The {\tt CheckerBoard} class.}
+{fig-checkerclass}
 \end{figure}
 
 The {\tt drawBoard()} method uses a nested {\tt for} loop to draw an
@@ -1783,9 +1778,8 @@ public class Average {                            // Console I/O
 \end{lstlisting}
 \end{jjjlistingleft}
 \jjjprogstop{A program to compute average grade using a {\tt while}
-structure.
-\label{fig-avg}
-\label{pg-fig-avg}}
+structure.}
+{fig-avg}
 \end{figure}
 
 One final point about this program is to note the care taken in the
@@ -1837,9 +1831,8 @@ while the grade < 0 or grade > 100 and grade != 9999
 
 \noindent Note here that initialization and updating of the loop
 \marginfig{chptr06/6f6.eps}{Do-while is a good structure for 
-the data validation algorithm.
-\label{fig-datavalid}
-\label{pg-fig-datavalid}}
+the data validation algorithm.}
+{fig-datavalid}
 variable are performed by the same statement. This is acceptable
 because we must update the value of {\tt grade} on each iteration {\it
 before} checking its validity. Note also that for this problem the
@@ -1905,9 +1898,8 @@ public class Validate {                            // Console input
 \end{lstlisting}
 \end{jjjlistingleft}
 \jjjprogstop{A program to compute average grade using a {\tt while}
-structure. This version validates the user's input.
-\label{fig-validate}
-\label{pg-fig-validate}}
+structure. This version validates the user's input.}
+{fig-validate}
 \end{figure}
 
 \section{Principles of Loop Design}
@@ -2148,9 +2140,9 @@ exactly the behavior of the multiway if-else selection structure
 \begin{figure}[hbt]
 \figa{chptr03/switchfl.eps}{Flowchart of the multiway switch structure.
 Note that because of the {\tt break} statement, one and only one case is
-executed.
-\label{fig-multiwayswitch}
-\label{pg-fig-multiwayswitch}}
+executed.}
+{fig-multiwayswitch}
+
 \end{figure}
 
 
@@ -2282,9 +2274,9 @@ in a program (Fig.~\ref{fig-controlstructs}):
 \begin{figure}[hbt]
 \figa{chptr06/6f10.eps}{Flowcharts of the four types of
 control structures. Each small rectangle represents a single executable
-statement.\index{control structures}
-\label{fig-controlstructs}
-\label{pg-fig-controlstructs}}
+statement.\index{control structures}}
+{fig-controlstructs}
+
 \end{figure}
 
 \begin{BL}
@@ -2430,9 +2422,8 @@ and the postcondition is that {\it k} should equal 1 on line
   } // factorial()
 \end{lstlisting}
 \end{jjjlisting}
-\jjjprogstop{Using pre- and postconditions to document code.
-\label{fig-preposts}
-\label{pg-fig-preposts}}
+\jjjprogstop{Using pre- and postconditions to document code.}
+{fig-preposts}
 \end{figure}
 
 In the first example, we use pre- and postconditions to define
