+++ texfiles/a.tex
@@ -389,8 +389,8 @@ public class Example {
 } // Example
 \end{lstlisting}
 \end{jjjlisting}
-\jjjprogstop{A sample Java source file.
-\label{fig-example1}\label{pg-fig-example1}}
+\jjjprogstop{A sample Java source file.}
+{fig-example1}
 \end{figure}
