+++ texfiles/f.tex
@@ -93,8 +93,8 @@ public class ConverterUser {
 } // ConverterUser
 \end{lstlisting}
 \end{jjjlistingleft}
-\jjjprogstop{A Java application containing a top-level nested class.
-\label{fig-converter}\label{pg-fig-converter}}
+\jjjprogstop{A Java application containing a top-level nested class. }
+{fig-converter}
 \end{figure}
 metric conversions.  The outer {\tt Converter} class serves as a
 container for the inner classes, {\tt Distance} and {\tt Weight},
@@ -235,9 +235,9 @@ public class ConverterFrame extends JFrame {
 } // ConverterFrame
 \end{lstlisting}
 \end{jjjlistingleft}
-\jjjprogstop{The use of a local class as an {\tt ActionListener} adapter.
-\label{fig-converterframe}\label{pg-fig-converterframe}}\end{figure}
-
+\jjjprogstop{The use of a local class as an {\tt ActionListener}
+  adapter.} {fig-converterframe}
+\end{figure}
 \noindent As we have seen, Java's event-handling model uses predefined
 interfaces, such as the {\tt ActionListener} interface, to handle
 events.  When a separate class is defined to implement an interface, it
