+++ texfiles/g.tex
@@ -223,8 +223,8 @@ public class Calendar {
 } // Calendar
 \end{lstlisting}
 \end{jjjlisting}
-\jjjprogstop{A Calendar class using the {\tt Season}.
-\label{fig-calendar}\label{pg-fig-calendar}}
+\jjjprogstop{A Calendar class using the {\tt Season}.}
+{fig-calendar}
 \end{figure}
 Note that the parameter of {\tt startDate()} is of type {\tt Season}
 and so the Java compiler can check that call to this method have
