+++ texfiles/jjj-os.tex
@@ -1,1 +1,1 @@
-% book Morelli% **********% ********** How to generate a PDF.% 0) makeindex jjj-os% 1) latex jjj-os% 2) dvips -f jjj-os.dvi > jjj-os.ps% 3) ps2pdf jjj-os.ps % **********% **********\documentclass{book}\input 3E.sty\input 3Enewstuff.sty%\makeindex%\includeonly{0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,a,b,c,d,e,f,g}%%\includeonly{a,b,c,d,e,f,g}%еееееее%%\includeonly{7.strings}\setcounter{page}{307}\usepackage{bookmark}\usepackage[]{hyperref}\hypersetup{    pdftitle={Java, Java, Java},    pdfauthor={Ralph Morelli and Ralph Walde},    pdfsubject={Java},    pdfkeywords={Computer Science},    bookmarksnumbered=true,         bookmarksopen=true,             bookmarksopenlevel=1,           colorlinks=true,                pdfstartview=Fit,               pdfpagemode=UseOutlines,    % this is the option you were lookin for    pdfpagelayout=TwoPageRight}\begin{document}\begin{titlepage}\title{\Huge{Java, Java, Java} \\ \huge {Object-Oriented Problem Solving} \\ \LARGE {Third Edition}}\author{\huge R. Morelli and R. Walde\\ \LARGE Trinity College\\ \LARGE Hartford, CT}\date{\LARGE \today}\maketitle\newpage\thispagestyle{empty}\vspace*{20pc}\hspace*{-3pc}This work is licensed under a \\Creative Commons Attribution 4.0 International License (CC BY 4.0).\\\\This book was previously published by \\Pearson Education, Inc.\end{titlepage}\frontmatter\setcounter{tocdepth}{1}\include{preface-os}\include{preface}%%%\include{foreword}\tableofcontents\mainmatter\setcounter{page}{1}\include{0.intro}      %% clean 4/9/05, spell-checked 4/9/05\include{1.programs}   %% clean 4/9/05, spell-checked 4/9/05\include{2.objects}    %% clean 4/10/05, spell-checked 4/10/05\include{3.methods}    %% clean 4/11/05, spell-checked 4/11/05\include{4.uis}        %% clean 4/11/05, spell-checked 4/11/05\include{5.data}       %% clean 4/13/05, spell-checked 4/13/05\include{6.loops}      %% clean 4/14/05, spell-checked 4/14/05\include{7.strings}    %% clean 4/14/05, spell-checked 4/14/05\include{8.oop}        %% clean 4/15/05, spell-checked 4/15/05\include{9.arrays}     %% clean, spell-checked 4/16/05\include{10.exceptions}     %% clean 4/4/05, spell-checked 4/7/05\include{11.files}          %% clean 4/5/05, spell-checked 4/5/05\include{12.recursion}     %% clean 4/6/05, spell-checked 4/7/05\include{13.guis}          %% clean 4/7/05, spell-checked 4/7/05\include{14.threads}       %% clean 4/7/05, spell-checked 4/7/05\include{15.sockets}        %% clean 4/7/05, spell-checked 4/7/05\include{16.datastructs}    %% clean 4/9/05, spell-checked 4/9/05\begin{appendix}\include{a}\include{b}\include{c}\include{d}\include{e}\include{f}\include{g}\include{h}\end{appendix}\printindex%%\endinput\end{document}
+% book Morelli% **********% ********** How to generate a PDF.% 0) makeindex jjj-os% 1) latex jjj-os% 2) dvips -f jjj-os.dvi > jjj-os.ps% 3) ps2pdf jjj-os.ps % **********% **********\documentclass{book}\input 3E.sty\input 3Enewstuff.sty%\makeindex%\includeonly{0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,a,b,c,d,e,f,g}%%\includeonly{a,b,c,d,e,f,g}%еееееее%%\includeonly{7.strings}\setcounter{page}{307}\usepackage{bookmark}\usepackage[]{hyperref}\usepackage{caption}\hypersetup{    pdftitle={Java, Java, Java},    pdfauthor={Ralph Morelli and Ralph Walde},    pdfsubject={Java},    pdfkeywords={Computer Science},    bookmarksnumbered=true,         bookmarksopen=true,             bookmarksopenlevel=1,           colorlinks=true,                pdfstartview=Fit,               pdfpagemode=UseOutlines,    % this is the option you were lookin for    pdfpagelayout=TwoPageRight}\begin{document}\begin{titlepage}\title{\Huge{Java, Java, Java} \\ \huge {Object-Oriented Problem Solving} \\ \LARGE {Third Edition}}\author{\huge R. Morelli and R. Walde\\ \LARGE Trinity College\\ \LARGE Hartford, CT}\date{\LARGE \today}\maketitle\newpage\thispagestyle{empty}\vspace*{20pc}\hspace*{-3pc}This work is licensed under a \\Creative Commons Attribution 4.0 International License (CC BY 4.0).\\\\This book was previously published by \\Pearson Education, Inc.\end{titlepage}\frontmatter\setcounter{tocdepth}{1}\include{preface-os}\include{preface}%%%\include{foreword}\tableofcontents\mainmatter\setcounter{page}{1}\include{0.intro}      %% clean 4/9/05, spell-checked 4/9/05\include{1.programs}   %% clean 4/9/05, spell-checked 4/9/05\include{2.objects}    %% clean 4/10/05, spell-checked 4/10/05\include{3.methods}    %% clean 4/11/05, spell-checked 4/11/05\include{4.uis}        %% clean 4/11/05, spell-checked 4/11/05\include{5.data}       %% clean 4/13/05, spell-checked 4/13/05\include{6.loops}      %% clean 4/14/05, spell-checked 4/14/05\include{7.strings}    %% clean 4/14/05, spell-checked 4/14/05\include{8.oop}        %% clean 4/15/05, spell-checked 4/15/05\include{9.arrays}     %% clean, spell-checked 4/16/05\include{10.exceptions}     %% clean 4/4/05, spell-checked 4/7/05\include{11.files}          %% clean 4/5/05, spell-checked 4/5/05\include{12.recursion}     %% clean 4/6/05, spell-checked 4/7/05\include{13.guis}          %% clean 4/7/05, spell-checked 4/7/05\include{14.threads}       %% clean 4/7/05, spell-checked 4/7/05\include{15.sockets}        %% clean 4/7/05, spell-checked 4/7/05\include{16.datastructs}    %% clean 4/9/05, spell-checked 4/9/05\begin{appendix}\include{a}\include{b}\include{c}\include{d}\include{e}\include{f}\include{g}\include{h}\end{appendix}\printindex%%\endinput\end{document}
